// src/post/gateways/path-types.ts
export enum PathType {
  GET_POSTS = 'get-posts',
  GET_POST = 'get-post',
  CREATE = 'create',
  DELETE = 'delete',
  COMMENT_POST = 'comment-post',
  REMOVE_COMMENT_POST = 'remove-comment-post',
  GET_SHOWING_POST = 'get-showing-post',
  TOGGLE_LIKE_POST = 'toggle-like-post',
  TOGGLE_LIKE_COMMENT = 'toggle-like-comment',
}